Reorg DWARF-related binary provider plugins #1167
Use own module to avoid any name collision
Provide proper path to binary
