+++ plugins/scancode-dwarfdump-macosx_10_9_intel/setup.cfg
