+++ plugins/scancode-dwarfdump-manylinux1_x86_64/.gitignore
