+++ plugins/scancode-dwarfdump-manylinux1_x86_64/src/scancode_dwarfdump/bin/dwarfdump2
