+++ plugins/scancode-dwarfdump-win32/.gitignore
