+++ plugins/scancode-dwarfdump-win32/MANIFEST.in
