+++ plugins/scancode-dwarfdump-win32/setup.cfg
