+++ plugins/scancode-dwarfdump-win_amd64/.gitignore
