+++ plugins/scancode-dwarfdump-win_amd64/setup.cfg
